third_party/WebKit/LayoutTests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/scroll-snap-type-expected.txt

This is a testharness.js-based test.
PASS Can set 'scroll-snap-type' to CSS-wide keywords
PASS Can set 'scroll-snap-type' to var() references
PASS Can set 'scroll-snap-type' to the 'none' keyword
FAIL Can set 'scroll-snap-type' to the 'x' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'scroll-snap-type' to the 'y' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'scroll-snap-type' to the 'block' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'scroll-snap-type' to the 'inline' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
FAIL Can set 'scroll-snap-type' to the 'both' keyword assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Setting 'scroll-snap-type' to a length throws TypeError
PASS Setting 'scroll-snap-type' to a percent throws TypeError
PASS Setting 'scroll-snap-type' to a time throws TypeError
PASS Setting 'scroll-snap-type' to an angle throws TypeError
PASS Setting 'scroll-snap-type' to a flexible length throws TypeError
PASS Setting 'scroll-snap-type' to a number throws TypeError
PASS Setting 'scroll-snap-type' to a position throws TypeError
PASS Setting 'scroll-snap-type' to a URL throws TypeError
PASS Setting 'scroll-snap-type' to a transform throws TypeError
PASS 'scroll-snap-type' does not supported 'x mandatory'
PASS 'scroll-snap-type' does not supported 'inline proximity'
Harness: the test ran to completion.
